refactored front-end and added event data clear
