+++ src/main/resources/templates/editSoldierData.html
@@ -4,7 +4,7 @@
         <meta name='viewport' content='width=device-width, initial-scale=1.0'>
         <link rel="stylesheet" href="main.css">
         <script type='module'>
-            import * as Controller from '/main.js';
+            import * as Controller from '/EditSoldierViewController.js';
             window.onload = Controller.editSoldierDataViewOnLoad;
             document.getElementById('returnToIndexButton').addEventListener('click', Controller.getHomePageViewController);
             document.getElementById('eventSelector').addEventListener('change', Controller.eventInputController);
