+++ src/main/resources/templates/index.html
@@ -4,7 +4,7 @@
         <meta name='viewport' content='width=device-width, initial-scale=1.0'>
         <link rel="stylesheet" href="main.css">
         <script type='module'>
-            import * as Controller from '/main.js';
+            import * as Controller from '/IndexViewController.js';
             window.onload = Controller.indexOnLoad;
             document.getElementById('createNewSoldierButton').addEventListener('click', Controller.createNewSoldierController);
             document.getElementById('editSoldierDataViewButton').addEventListener('click', Controller.showEditSoldierDataViewController);
